+++ packages/ckeditor5-list/src/index.ts
@@ -26,9 +26,9 @@ export type { default as ListReversedCommand } from './listproperties/listrevers
 export type { default as ListStartCommand } from './listproperties/liststartcommand.js';
 export type { default as ListStyleCommand } from './listproperties/liststylecommand.js';
 
-// ListItemFormatting.
-export { default as ListItemFormatting } from './listitemformatting.js';
-export { default as ListItemFontFamilyIntegration } from './listitemformatting/listitemfontfamilyintegration.js';
+// ListFormatting.
+export { default as ListFormatting } from './listformatting.js';
+export { default as ListItemFontFamilyIntegration } from './listformatting/listitemfontfamilyintegration.js';
 
 // TodoList/
 export { default as TodoList } from './todolist.js';
