+++ packages/ckeditor5-list/tests/list/listformatting.js
@@ -1,0 +1,50 @@
+/**
+ * @license Copyright (c) 2003-2025, CKSource Holding sp. z o.o. All rights reserved.
+ * For licensing, see LICENSE.md or https://ckeditor.com/legal/ckeditor-licensing-options
+ */
+
+import Paragraph from '@ckeditor/ckeditor5-paragraph/src/paragraph.js';
+import VirtualTestEditor from '@ckeditor/ckeditor5-core/tests/_utils/virtualtesteditor.js';
+import testUtils from '@ckeditor/ckeditor5-core/tests/_utils/utils.js';
+// import { getData as getModelData, setData as setModelData } from '@ckeditor/ckeditor5-engine/src/dev-utils/model.js';
+// import { getData as getViewData } from '@ckeditor/ckeditor5-engine/src/dev-utils/view.js';
+
+import stubUid from '@ckeditor/ckeditor5-list/tests/list/_utils/uid.js';
+
+import ListFormatting from '../../src/listformatting.js';
+
+describe( 'ListFormatting', () => {
+	let editor;
+
+	testUtils.createSinonSandbox();
+
+	beforeEach( async () => {
+		editor = await VirtualTestEditor.create( {
+			plugins: [ ListFormatting, Paragraph ]
+		} );
+
+		// model = editor.model;
+
+		stubUid();
+	} );
+
+	afterEach( async () => {
+		await editor.destroy();
+	} );
+
+	it( 'should have pluginName', () => {
+		expect( ListFormatting.pluginName ).to.equal( 'ListFormatting' );
+	} );
+
+	it( 'should have `isOfficialPlugin` static flag set to `true`', () => {
+		expect( ListFormatting.isOfficialPlugin ).to.be.true;
+	} );
+
+	it( 'should have `isPremiumPlugin` static flag set to `false`', () => {
+		expect( ListFormatting.isPremiumPlugin ).to.be.false;
+	} );
+
+	it( 'should be loaded', () => {
+		expect( editor.plugins.get( ListFormatting ) ).to.be.instanceOf( ListFormatting );
+	} );
+} );
